--- a/crates/analyzer/src/db/queries/module.rs
+++ b/crates/analyzer/src/db/queries/module.rs
@@ -15,6 +15,7 @@ use fe_parser::ast;
 use fe_parser::node::Node;
 use indexmap::indexmap;
 use indexmap::map::{Entry, IndexMap};
+use smol_str::SmolStr;
 use std::collections::HashSet;
 use std::ops::Deref;
 use std::path::Path;
@@ -22,27 +23,24 @@ use std::rc::Rc;
 use strum::IntoEnumIterator;
 
 // Placeholder; someday std::prelude will be a proper module.
-fn std_prelude_items() -> IndexMap<String, Item> {
+fn std_prelude_items() -> IndexMap<SmolStr, Item> {
     let mut items = indexmap! {
-        "bool".to_string() => Item::Type(TypeDef::Primitive(types::Base::Bool)),
-        "address".to_string() => Item::Type(TypeDef::Primitive(types::Base::Address)),
+        SmolStr::new("bool") => Item::Type(TypeDef::Primitive(types::Base::Bool)),
+        SmolStr::new("address") => Item::Type(TypeDef::Primitive(types::Base::Address)),
     };
     items.extend(types::Integer::iter().map(|typ| {
         (
-            typ.as_ref().to_string(),
+            typ.as_ref().into(),
             Item::Type(TypeDef::Primitive(types::Base::Numeric(typ))),
         )
     }));
-    items.extend(
-        types::GenericType::iter().map(|typ| (typ.name().to_string(), Item::GenericType(typ))),
-    );
+    items.extend(types::GenericType::iter().map(|typ| (typ.name(), Item::GenericType(typ))));
     items.extend(
         builtins::GlobalFunction::iter()
-            .map(|fun| (fun.as_ref().to_string(), Item::BuiltinFunction(fun))),
+            .map(|fun| (fun.as_ref().into(), Item::BuiltinFunction(fun))),
     );
-    items.extend(
-        builtins::GlobalObject::iter().map(|obj| (obj.as_ref().to_string(), Item::Object(obj))),
-    );
+    items
+        .extend(builtins::GlobalObject::iter().map(|obj| (obj.as_ref().into(), Item::Object(obj))));
     items
 }
 
@@ -60,7 +58,7 @@ pub fn module_all_items(db: &dyn AnalyzerDb, module: ModuleId) -> Rc<Vec<Item>> 
             ))),
             ast::ModuleStmt::Contract(node) => Some(Item::Type(TypeDef::Contract(
                 db.intern_contract(Rc::new(Contract {
-                    name: node.name().to_string(),
+                    name: node.name().into(),
                     ast: node.clone(),
                     module,
                 })),
@@ -94,7 +92,7 @@ pub fn module_all_items(db: &dyn AnalyzerDb, module: ModuleId) -> Rc<Vec<Item>> 
 pub fn module_item_map(
     db: &dyn AnalyzerDb,
     module: ModuleId,
-) -> Analysis<Rc<IndexMap<String, Item>>> {
+) -> Analysis<Rc<IndexMap<SmolStr, Item>>> {
     let builtin_items = std_prelude_items();
     let sub_modules = module
         .sub_modules(db)
@@ -104,7 +102,7 @@ pub fn module_item_map(
     let used_items = db.module_used_item_map(module);
     let mut diagnostics = (*used_items.diagnostics).clone();
 
-    let mut map = IndexMap::<String, Item>::new();
+    let mut map = IndexMap::<SmolStr, Item>::new();
 
     for item in module.all_items(db).iter() {
         let item_name = item.name(db);
@@ -232,7 +230,7 @@ pub fn module_constant_type(
 pub fn module_used_item_map(
     db: &dyn AnalyzerDb,
     module: ModuleId,
-) -> Analysis<Rc<IndexMap<String, (Span, Item)>>> {
+) -> Analysis<Rc<IndexMap<SmolStr, (Span, Item)>>> {
     let mut diagnostics = vec![];
 
     let ast::Module { body } = &module.data(db).ast;
@@ -296,7 +294,7 @@ pub fn module_resolve_use_tree(
     db: &dyn AnalyzerDb,
     module: ModuleId,
     tree: Node<ast::UseTree>,
-) -> Analysis<Rc<IndexMap<String, (Span, Item)>>> {
+) -> Analysis<Rc<IndexMap<SmolStr, (Span, Item)>>> {
     let mut diagnostics = vec![];
 
     match &tree.kind {
@@ -421,7 +419,7 @@ pub fn module_parent_module(db: &dyn AnalyzerDb, module: ModuleId) -> Option<Mod
 pub fn module_adjacent_modules(
     db: &dyn AnalyzerDb,
     module: ModuleId,
-) -> Rc<IndexMap<String, ModuleId>> {
+) -> Rc<IndexMap<SmolStr, ModuleId>> {
     if let Some(parent) = module.parent_module(db) {
         parent.sub_modules(db)
     } else {
@@ -429,7 +427,10 @@ pub fn module_adjacent_modules(
     }
 }
 
-pub fn module_sub_modules(db: &dyn AnalyzerDb, module: ModuleId) -> Rc<IndexMap<String, ModuleId>> {
+pub fn module_sub_modules(
+    db: &dyn AnalyzerDb,
+    module: ModuleId,
+) -> Rc<IndexMap<SmolStr, ModuleId>> {
     match module.context(db) {
         ModuleContext::Ingot(ingot) => {
             let all_modules = ingot.all_modules(db);
@@ -439,10 +440,10 @@ pub fn module_sub_modules(db: &dyn AnalyzerDb, module: ModuleId) -> Rc<IndexMap<
                     let sub_modules = all_modules
                         .iter()
                         .filter(|module_id| {
-                            Path::new(&module_id.ingot_path(db))
+                            Path::new(module_id.ingot_path(db).as_str())
                                 .parent()
                                 .expect("module file in ingot does not have parent path")
-                                == Path::new(&dir_path)
+                                == Path::new(dir_path.as_str())
                         })
                         .map(|module_id| (module_id.name(db), *module_id))
                         .collect::<IndexMap<_, _>>();