… Cargo.lock
… crates/analyzer/Cargo.toml
… crates/analyzer/src/namespace/items.rs
… crates/analyzer/src/traversal/expressions.rs
… crates/analyzer/tests/analysis.rs
… crates/analyzer/tests/errors.rs
… crates/driver/src/lib.rs
… crates/lowering/src/ast_utils.rs
… crates/lowering/src/mappers/expressions.rs
… crates/lowering/src/mappers/functions.rs
… crates/lowering/src/mappers/module.rs
… crates/lowering/src/names.rs
… crates/lowering/tests/lowering.rs
… crates/parser/Cargo.toml
… crates/parser/benches/bench.rs
… crates/parser/src/ast.rs
… crates/parser/src/grammar/contracts.rs
… crates/parser/src/grammar/expressions.rs
… crates/parser/src/grammar/functions.rs
… crates/parser/src/grammar/module.rs
… crates/parser/src/grammar/types.rs
… crates/parser/src/lexer/token.rs
… crates/parser/src/node.rs
… crates/test-utils/Cargo.toml
… crates/yulgen/src/db/queries/contracts.rs
… crates/yulgen/src/mappers/expressions.rs
… crates/yulgen/src/mappers/module.rs
… crates/yulgen/src/runtime/abi_dispatcher.rs
… crates/yulgen/tests/yulgen.rs