### What was wrong?

```
parse_uniswap           time:   [5.5317 ms 5.5724 ms 5.6175 ms]                          
clone_ast               time:   [146.20 us 147.06 us 148.00 us]                      
analyze uniswap         time:   [1.3687 ms 1.3788 ms 1.3894 ms]                             
```

### How was it fixed?

```
parse_uniswap           time:   [252.10 us 254.47 us 256.97 us]                          
                        change: [-95.455% -95.391% -95.324%] (p = 0.00 < 0.05)

clone_ast               time:   [78.725 us 79.315 us 79.954 us]                      
                        change: [-47.624% -46.448% -45.283%] (p = 0.00 < 0.05)

analyze uniswap         time:   [1.1384 ms 1.1481 ms 1.1590 ms]                             
                        change: [-18.654% -15.780% -12.782%] (p = 0.00 < 0.05)
```

Most of the parsing speedup came from changing `NodeId` from `Uuid::new_v4()` to a u32 (generated by incrementing a static atomic). 5.5ms -> 320us. Replacing Strings in the AST with SmolStr brought it down to 255us, and made AST cloning cheaper.


### To-Do

[//]: # (Stay ahead of things, add list items here!)
- [ ] OPTIONAL: Update [Spec](https://github.com/ethereum/fe/blob/master/docs/src/spec/index.md) if applicable
- [ ] Add entry to the [release notes](https://github.com/ethereum/fe/blob/master/newsfragments/README.md) (may forgo for trivial changes)

- [ ] Clean up commit history